refactor!: nest RequestControls under Config and document limit/schema ownership
